+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.10.0] - 2022-08-23
+
+### Added
+
+-   Added customSkinInfo method and its corresponding types
+-   Added renders beatmapsetid query
+
+### Changed
+
+-   Dependencies update
+
 ## [3.9.0] - 2022-07-09
 
 ### Added